Change variable declaration to `alias` in generate-env.js
When declared as a constant, the envvar can't be used in string concatenation (string interpolation). For example, say you have an authenticated endpoint. When envvars are constants, you can't use it when creating an `alias`.

If this PR is merged, this fixes the issue and allows envvars to be used in string interpolation scenarios.

fixes #15017